% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestTransformeR.r
\name{BestTransformeR}
\alias{BestTransformeR}
\title{Best data transformation identification function}
\usage{
BestTransformeR(x, y, type)
}
\arguments{
\item{x}{Non-Normalized input variable that needs to be transformed}

\item{y}{Response Variable}

\item{type}{Character string specifying the type of data transformation to use. Value "all" considers all the data transformation techniques. "Non-growth" – considers all the data transformation techniques other than growth models. "growth" - considers all the growth model techniques}
}
\description{
This function identifies the best data transformation for one variable to maximize the strength of its relationship with another variable. The different transformation techniques compared in this function are MinMax, Zscore, Square, Cube, Square Root, Cube Root,Box-Cox, Exponential & Arcsine and Growth Models: Gompertz, Chapman Richards, Weibull & Morgan-Mercer-Flodin. The best transformation technique is identified based on maximum R Squared Value.
}
\section{Warning}{

Normalized x variable shouldn’t be passed to the function as normalization happens within the function for relevant transformation techniques; Do not pass the x variable with zero value to this function.
}

\examples{
BestTranform(input_variable,response_variable,'growth')
BestTranform(input_variable,response_variable,'all')
BestTranform(input_variable,response_variable,'non-growth')
}
\keyword{Technique}
\keyword{Transformation}
