% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalOutliers.r
\name{CriticalOutliers}
\alias{CriticalOutliers}
\title{Critical Outlier Identification Function}
\usage{
CriticalOutliers(x, y, n, m)
}
\arguments{
\item{x}{Predictor variable}

\item{y}{Response Variable}

\item{n}{Number of critical outliers to be identified
m     Residual Error threshold. Takes values greater than 0, ideally between 0 and 1}
}
\description{
This function identifies the critical outliers in the predictor variable. Critical outliers are defined as data points which would maximize the accuracy of the simple linear regression model when handled
}
\examples{
CriticalOutliers(predictor_variable,response_variable,4,0.1)
}
\keyword{Critical}
\keyword{Identification}
\keyword{Outlier}
