% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOutliers.r
\name{FindOutliers}
\alias{FindOutliers}
\title{Outlier identification Method}
\usage{
FindOutliers(x, y, n)
}
\arguments{
\item{x}{Predictor variable}

\item{y}{Response Variable}

\item{n}{Residual Error threshold. Takes values greater than 0, ideally between 0 and 1}
}
\description{
This function identifies the outliers in the predictor variable based on residual error from a simple linear regression model.
}
\examples{
FindOutliers(predictor_variable,response_variable,0.1)
}
\keyword{Detection}
\keyword{Outlier}
