% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andThen.R
\name{Transformation.andThen1}
\alias{Transformation.andThen1}
\title{Chain Two Transformations}
\usage{
Transformation.andThen1(before, after)
}
\arguments{
\item{before}{the first transformation}

\item{after}{the transformation to be applied after \code{before}}
}
\value{
the corresponding \code{Transformation}
}
\description{
Create a new \code{\link{Transformation}} object resulting from
  the application of a \code{Transformation} to another one
}
\seealso{
\code{\link{Transformation.andThen2}}
}
