% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andThen.R
\name{Transformation.andThen2}
\alias{Transformation.andThen2}
\title{Chain a Transformation given by Two Functions and another
  Transformation}
\usage{
Transformation.andThen2(before, after.forward, after.backward)
}
\arguments{
\item{before}{the transformation to be applied first}

\item{after.forward}{the forward function to be applied to the result of the
forward function of \code{before}}

\item{after.backward}{the corresponding backward function}
}
\value{
the corresponding new \code{Transformation} object
}
\description{
Create a new \code{Transformation} object resulting from the
  application of a new function \code{fwd} to the output of this object's
  \code{forward} function.
}
\seealso{
\code{\link{Transformation.andThen1}}
}
