% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyAll.R
\name{Transformation.applyAll}
\alias{Transformation.applyAll}
\title{Apply All the Provided Transformations}
\usage{
Transformation.applyAll(data, transformations, addIdentity = TRUE)
}
\arguments{
\item{data}{the data}

\item{transformations}{the transformations}

\item{addIdentity}{should an identity transformation result be created as
well?}
}
\value{
a list with the transformation results, or \code{NULL} if no
  transformation succeeded
}
\description{
Provide all the provided transformation functions.
}
