% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyAll2D.R
\name{Transformation.applyAll2D}
\alias{Transformation.applyAll2D}
\title{Apply All the Provided Transformations to a Two-Dimensional Dataset}
\usage{
Transformation.applyAll2D(x, y, x.transformations,
  y.transformations = x.transformations, x.addIdentity = FALSE,
  y.addIdentity = x.addIdentity)
}
\arguments{
\item{x}{the \code{x} data}

\item{y}{the \code{y} data}

\item{x.transformations}{the transformations to be applied to the \code{x}
data}

\item{y.transformations}{the transformations to be applied to the \code{y}
data, by default the same as \code{x.transformations}}

\item{x.addIdentity}{should an identity transformation result be created as
well for \code{x}? (\code{FALSE} by default)}

\item{y.addIdentity}{should an identity transformation result be created as
well for \code{y}? (by default, the same as \code{x.addIdentity})}
}
\description{
Provide all the provided transformation functions to the \code{x} and \code{y} data.
}
