% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default2D.R
\name{Transformation.applyDefault2D}
\alias{Transformation.applyDefault2D}
\title{Apply All Default Transformations to a Two-Dimensional Dataset}
\usage{
Transformation.applyDefault2D(x, y, addIdentity = FALSE)
}
\arguments{
\item{x}{the \code{x} data}

\item{y}{the \code{y} data}

\item{addIdentity}{should an identity transformation result be created as
well? (by default \code{FALSE})}
}
\value{
a list with the transformation results, or \code{NULL} if no
  transformation succeeded
}
\description{
Apply all default transformations to a dataset.
}
\seealso{
\code{\link{Transformation.applyAll2D}}
}
