% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{Transformation.identity}
\alias{Transformation.identity}
\title{Create an Identity \code{\link{TransformedData}} Instance}
\usage{
Transformation.identity(data)
}
\arguments{
\item{data}{the data}
}
\value{
the identity transformed data
}
\description{
The resulting data will be the original data, i.e., an identity
  transformation.
}
\examples{
Transformation.identity(c(1,2,3))
# An object of class "TransformedData"
# Slot "transformation":
#   An object of class "Transformation"
# Slot "forward":
#   function (x)
#     x
# <bytecode: 0x3fe24f0>
#   <environment: namespace:base>
#
#   Slot "backward":
#   function (x)
#     x
# <bytecode: 0x3fe25d0>
#   <environment: namespace:base>
#
#   Slot "complexity":
#   [1] 0
#
# Slot "data":
#   [1] 1 2 3
}
\seealso{
\code{\link{TransformedData.new}}
}
