% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity2D.R
\name{Transformation.identity2D}
\alias{Transformation.identity2D}
\title{Instantiate a new \code{\link{TransformedData2D}} Object}
\usage{
Transformation.identity2D(x, y)
}
\arguments{
\item{x}{the x data vector}

\item{y}{the y data vector}
}
\value{
a new instance of \code{\link{TransformedData2D}}
}
\description{
Always use this function for instantiating
  \code{\link{TransformedData2D}}
}
