% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logScale.R
\name{Transformation.logNegated}
\alias{Transformation.logNegated}
\title{Log-Transform and then Normalize and Negate a Data Vector}
\usage{
Transformation.logNegated(data)
}
\arguments{
\item{data}{the data vector}
}
\value{
a \code{\link{TransformedData}} instance corresponding to the
  logarithmic mapping of \code{data} into \code{[1, 0]}, or \code{NULL} if
  such a mapping is not possible, e.g., if \code{data} contained non-finite
  values
}
\description{
A data vector is logarithmically scaled and then normalized and
  negated, i.e., all of its components are logarithmically mapped into the
  interval \code{[1, 0]}.
}
\seealso{
\code{\link{Transformation.apply}}
}
