% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePositive.R
\name{Transformation.makePositive}
\alias{Transformation.makePositive}
\title{Create a Transformation which Shifts all Values of a Data Array into
  the Positive Zone}
\usage{
Transformation.makePositive(data, zeroAllowed = TRUE)
}
\arguments{
\item{data}{the data vector}

\item{zeroAllowed}{is it ok if the transformation of the data vector would
produce 0 values?}
}
\value{
a \code{\link{Transformation}} which would transform the data vector
  into the positive domain, or \code{NULL} if no such transformation can be
  provided
}
\description{
A transformation is created which could be used to shift all
  elements of a data vector into the positive range. This is a utility
  function which can be applied before log-scaling or \code{1/x} scaling of
  data. The transformation will be a linear shift and not involve scaling
  or squaring.
}
\examples{
data <- c(-1, 0, 0.1, 2)
trafo <- Transformation.makePositive(data, zeroAllowed=TRUE)
trafo@forward(data)
# [1] 0.0 1.0 1.1 3.0
trafo@backward(trafo@forward(data))
# [1] -1.0  0.0  0.1  2.0
trafo@forward
# function (x)
# x + 1
# <environment: 0x4bee3c0>
trafo@complexity
# [1] 1
trafo <- Transformation.makePositive(data, zeroAllowed=FALSE)
trafo@forward(data)
# [1] 0.0625 1.0625 1.1625 3.0625
trafo@backward(trafo@forward(data))
# [1] -1.0  0.0  0.1  2.0
trafo@complexity
# [1] 2
}
