% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{Transformation.mapIntervals}
\alias{Transformation.mapIntervals}
\title{Create a Bijective Mapping between two Intervals}
\usage{
Transformation.mapIntervals(domainMin, domainMax, imageMin = 0,
  imageMax = 1)
}
\arguments{
\item{domainMin}{the minimum of the original interval}

\item{domainMax}{the maximum of the original interval}

\item{imageMin}{the minimum of the goal interval}

\item{imageMax}{the maximum of the goal interval}
}
\value{
a bijection between \code{[d, domainMax]} and \code{[0,1]}
}
\description{
Create a bijective mapping between the two real intervals
  \code{[domainMin, domainMax]} and \code{[imageMin, imageMax]}. The returned
  mapping will map \code{domainMin} to \code{imageMin} and \code{domainMax}
  to \code{imageMax} and vice versa.

  Notice that \code{domainMin>domainMax} as well as \code{imageMi >imageMax}
  are both permissible, in which case smaller input values may map to larger
  output values.

  Furthermore, although we specify domain and image boundaries, the returned
  mapping will not have any boundaries. It is entirely allowed to map values
  smaller than \code{domainMin} or larger than \code{domainMax}. These may
  then land outside of \code{[imageMin, imageMax]}, though.
}
\examples{
trafo <- Transformation.mapIntervals(-1, 1, 0, 2)
trafo
# An object of class "Transformation"
# Slot "forward":
# function (x)
# x + 1
# <environment: 0x35519c0>
#
# Slot "backward":
# function (x)
# x - 1
# <environment: 0x35519c0>
trafo <- Transformation.mapIntervals(0.2, 12, -0.1, 2)
trafo
# An object of class "Transformation"
# Slot "forward":
# function (x)
# (2.1 * ((x - 0.2)/11.8)) - 0.1
# <environment: 0x3c877f0>
#
# Slot "backward":
# function (x)
# 0.2 + (11.8 * ((x + 0.1)/2.1))
# <environment: 0x3c877f0>
}
\seealso{
\code{\link{Transformation.normalizeInterval}}
}
