% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\name{Transformation.new}
\alias{Transformation.new}
\title{Construct a new Transformation}
\usage{
Transformation.new(forward, backward)
}
\arguments{
\item{forward}{the forward mapping}

\item{backward}{the backward mapping}
}
\value{
the new \code{\link{Transformation}}
}
\description{
Always use this method instead of the constructor to build a new
  instance of \code{\link{Transformation}}, as it will canonicalize the
  functions used and ensure that all promises are resolved before returning
  the result.
}
