% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{Transformation.normalize}
\alias{Transformation.normalize}
\title{Normalize a Data Vector}
\usage{
Transformation.normalize(data)
}
\arguments{
\item{data}{the data vector}
}
\value{
a \code{\link{TransformedData}} instance corresponding to the linear
  mapping of \code{data} into \code{[0, 1]}, or \code{NULL} if such a mapping
  is not possible, e.g., if \code{data} contained non-finite values
}
\description{
A data vector is normalized, i.e., all of its components are
  linearly mapped into the interval \code{[0, 1]}.
}
\seealso{
\code{\link{Transformation.apply}}
}
