% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{Transformation.normalizeInterval}
\alias{Transformation.normalizeInterval}
\title{Create a Normalization Transformation}
\usage{
Transformation.normalizeInterval(domainMin, domainMax)
}
\arguments{
\item{domainMin}{the minimum of the original interval}

\item{domainMax}{the maximum of the original interval}
}
\value{
a bijection between \code{[domainMin, domainMax]} and \code{[0,1]}
}
\description{
Create a bijective mapping between the real interval
  \code{[domainMin, domainMax]} and \code{[0, 1]}. The returned mapping will
  map \code{domainMin} to \code{0} and \code{domainMax} to \code{1}. Notice
  that \code{domainMin>domainMax} is permissible, in which case smaller input
  values map to larger output values.
}
\examples{
Transformation.normalizeInterval(1, 3)
# An object of class "Transformation"
# Slot "forward":
# function (x)
# ((x - 1)/2)
# <environment: 0x3bf9960>
#
# Slot "backward":
# function (x)
# (x * 2) + 1
# <environment: 0x3bf9960>
}
\seealso{
\code{\link{Transformation.mapIntervals}}
}
