% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformedData.R
\docType{class}
\name{TransformedData-class}
\alias{TransformedData-class}
\alias{TransformedData}
\title{The Result of a Single Transformation Step}
\description{
A class for presenting the results of the transformation of a
  data column based on a \code{\link{Transformation}}
}
\section{Slots}{

\describe{
\item{\code{transformation}}{the \code{\link{Transformation}} between the original data and the transformed
data}

\item{\code{data}}{the transformed data}
}}

