% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformedData.R
\name{TransformedData.new}
\alias{TransformedData.new}
\title{Instantiate a new \code{\link{TransformedData}} Object}
\usage{
TransformedData.new(transformation, data)
}
\arguments{
\item{transformation}{the transformation}

\item{data}{the transformed data}
}
\value{
a new instance of \code{\link{TransformedData}}
}
\description{
Always use this function for instantiating
  \code{\link{TransformedData}}
}
