% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{TransformedData.select1D}
\alias{TransformedData.select1D}
\title{Select a Sub-Set of a \code{\link{TransformedData}} Instance}
\usage{
TransformedData.select1D(data, selection)
}
\arguments{
\item{data}{the \code{\link{TransformedData}} instance}

\item{selection}{the selection, or \code{NULL} to select all}
}
\value{
the corresponding subset
}
\description{
This method selects a subset of the provided instance of
  \code{\link{TransformedData}}
}
