% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformedData2D.R
\docType{class}
\name{TransformedData2D-class}
\alias{TransformedData2D-class}
\alias{TransformedData2D}
\title{A Dataset Consisting of \code{x} and \code{y} Data}
\description{
This class is the baseline for fitting models to \code{x}
and \code{y} data. It can encapsulate the data as well as
transformations that have been applied to it.
}
\section{Slots}{

\describe{
\item{\code{x}}{the x \code{\link{TransformedData}} instance}

\item{\code{y}}{the y \code{\link{TransformedData}} instance}
}}

