% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformedData2D.R
\name{TransformedData2D.new}
\alias{TransformedData2D.new}
\title{Instantiate a new \code{\link{TransformedData2D}} Object}
\usage{
TransformedData2D.new(x, y)
}
\arguments{
\item{x}{the x \code{\link{TransformedData}} instance}

\item{y}{the y \code{\link{TransformedData}} instance}
}
\value{
a new instance of \code{\link{TransformedData2D}}
}
\description{
Always use this function for instantiating
  \code{\link{TransformedData2D}}
}
