% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformation.R
\docType{class}
\name{Transformation-class}
\alias{Transformation-class}
\alias{Transformation}
\title{A Transformation, i.e. a Two-Way Mapping}
\description{
This class holds two mutually inverse functions for transforming
 data, i.e., allows us to transform data forward and backward. It should hold
 \code{b@backward(b@forward(x))==x} for a Transformation \code{b} and values
 \code{x} in the domain of the Transformation. At the same time,
 \code{b@forward(b@backward(y))==y} should hold for all \code{y} in the image
 of the Transformation.
}
\section{Slots}{

\describe{
\item{\code{forward}}{the foward transformation function}

\item{\code{backward}}{the inverse of the transformation function, i.e., the backwards
transformation}
}}

