% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity.R
\name{Transformation.identity}
\alias{Transformation.identity}
\title{Create an Identity \code{\link{TransformedData}} Instance}
\usage{
Transformation.identity(data)
}
\arguments{
\item{data}{the data}
}
\value{
the identity transformed data
}
\description{
The resulting data will be the original data, i.e., an identity
  transformation.
}
\seealso{
\code{\link{TransformedData.new}}
}
