% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logScale.R
\name{Transformation.logNegated}
\alias{Transformation.logNegated}
\title{Log-Transform and then Normalize and Negate a Data Vector}
\usage{
Transformation.logNegated(data)
}
\arguments{
\item{data}{the data vector}
}
\value{
a \code{\link{TransformedData}} instance corresponding to the
  logarithmic mapping of \code{data} into \code{[1, 0]}, or \code{NULL} if
  such a mapping is not possible, e.g., if \code{data} contained non-finite
  values
}
\description{
A data vector is logarithmically scaled and then normalized and
  negated, i.e., all of its components are logarithmically mapped into the
  interval \code{[1, 0]}.
}
\examples{
data <- c(-1, 0, 2, 6, 14, 30)
Transformation.logNegated(data)
# An object of class "TransformedData"
# Slot "transformation":
#   An object of class "Transformation"
# Slot "forward":
#   function (x)
#     ((log(x + 2) - 3.46573590279973)/-3.46573590279973)
# <environment: 0x6e9b180>
#
#   Slot "backward":
#   function (x)
#     exp(x = (x * -3.46573590279973) + 3.46573590279973) - 2
# <environment: 0x5481518>
#
#   Slot "complexity":
#   [1] 6
#
# Slot "data":
#   [1] 1.0 0.8 0.6 0.4 0.2 0.0
}
\seealso{
\code{\link{Transformation.apply}}
}
